finished cleaning the code up
